% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.dglm.R
\name{residuals.dglm}
\alias{residuals.dglm}
\title{residuals.dglm function}
\usage{
\method{residuals}{dglm}(object, ...)
}
\arguments{
\item{object}{The fitted dglm object whose residuals we want}

\item{...}{additional parameters to residuals.glm}
}
\value{
the residuals from the mean portion of the dglm object
}
\description{
This implements the 'residuals' generic for the dglm object
}
\author{
Robert W. Corty
}

